+++ examples/ui/ui.rs
@@ -21,7 +21,7 @@ fn main() {
 
     #[cfg(feature = "bevy_dev_tools")]
     {
-        app.add_plugins(bevy::dev_tools::debug_overlay::DebugUiPlugin)
+        app.add_plugins(bevy::dev_tools::ui_debug_overlay::DebugUiPlugin)
             .add_systems(Update, toggle_overlay);
     }
 
@@ -374,7 +374,7 @@ fn mouse_scroll(
 // The system that will enable/disable the debug outlines around the nodes
 fn toggle_overlay(
     input: Res<ButtonInput<KeyCode>>,
-    mut options: ResMut<bevy::dev_tools::debug_overlay::UiDebugOptions>,
+    mut options: ResMut<bevy::dev_tools::ui_debug_overlay::UiDebugOptions>,
 ) {
     info_once!("The debug outlines are enabled, press Space to turn them on/off");
     if input.just_pressed(KeyCode::Space) {
